+++ json_annotation/lib/src/json_serializable.dart
@@ -3,20 +3,34 @@
 // BSD-style license that can be found in the LICENSE file.
 
 class JsonSerializable {
+  // TODO(kevmoo): document these fields
   final bool createFactory;
   final bool createToJson;
 
-  /// Whether the generator should include the this field in the serialized
-  /// output, even if the value is `null`.
+  /// Whether the generator should include fields with `null` values in the
+  /// serialized output.
   final bool includeIfNull;
 
+  /// When `true` (the default), `null` values are handled gracefully when
+  /// serializing fields to JSON and when deserializing `null` and nonexistent
+  /// values from a JSON map.
+  ///
+  /// Setting to `false` eliminates `null` verification in the generated code,
+  /// which reduces the code size. Errors may be thrown at runtime if `null`
+  /// values are encountered, but the original class should also implement
+  /// `null` runtime validation if it's critical.
+  final bool nullable;
+
+  // TODO(kevmoo): document the constructor
   const JsonSerializable(
       {bool createFactory: true,
       bool createToJson: true,
-      bool includeIfNull: true})
+      bool includeIfNull: true,
+      bool nullable: true})
       : this.createFactory = createFactory ?? true,
         this.createToJson = createToJson ?? true,
-        this.includeIfNull = includeIfNull ?? true;
+        this.includeIfNull = includeIfNull ?? true,
+        this.nullable = nullable ?? true;
 }
 
 /// An annotation used to specify how a field is serialized.
@@ -27,15 +41,21 @@ class JsonKey {
   /// If `null`, the field name is used.
   final String name;
 
-  /// [true] if the generator should validate all values for `null` in
-  /// serialization code.
+  /// When `true`, `null` values are handled gracefully when
+  /// serializing the field to JSON and when deserializing `null` and
+  /// nonexistent values from a JSON map.
   ///
-  /// Setting to [false] eliminates `null` verification in generated code, but
-  /// does not prevent `null` values from being created. Annotated classes
-  /// must implement their own `null` validation.
+  /// Setting to `false` eliminates `null` verification in the generated code
+  /// for the annotated field, which reduces the code size. Errors may be thrown
+  /// at runtime if `null` values are encountered, but the original class should
+  /// also implement `null` runtime validation if it's critical.
+  ///
+  /// The default value, `null`, indicates that the behavior should be
+  /// acquired from the [JsonSerializable.nullable] annotation on the
+  /// enclosing class.
   final bool nullable;
 
-  /// [true] if the generator should include the this field in the serialized
+  /// `true` if the generator should include the this field in the serialized
   /// output, even if the value is `null`.
   ///
   /// The default value, `null`, indicates that the behavior should be
@@ -46,7 +66,5 @@ class JsonKey {
   /// Creates a new [JsonKey].
   ///
   /// Only required when the default behavior is not desired.
-  const JsonKey({this.name, bool nullable: true, bool includeIfNull})
-      : this.nullable = nullable ?? true,
-        this.includeIfNull = includeIfNull;
+  const JsonKey({this.name, this.nullable, this.includeIfNull});
 }