+++ json_serializable/test/test_files/kitchen_sink.non_nullable.g.dart
@@ -1,0 +1,114 @@
+// Copyright (c) 2017, the Dart project authors. Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of json_serializable.test.kitchen_sink_nullable;
+
+// **************************************************************************
+// Generator: JsonSerializableGenerator
+// **************************************************************************
+
+KitchenSinkNonNullable _$KitchenSinkNonNullableFromJson(
+        Map<String, dynamic> json) =>
+    new KitchenSinkNonNullable(
+        iterable: json['iterable'] as List,
+        dynamicIterable: json['dynamicIterable'] as List,
+        objectIterable: json['objectIterable'] as List,
+        intIterable: (json['intIterable'] as List).map((e) => e as int),
+        dateTimeIterable: (json['datetime-iterable'] as List)
+            .map((e) => DateTime.parse(e as String)))
+      ..dateTime = DateTime.parse(json['dateTime'] as String)
+      ..list = json['list'] as List
+      ..dynamicList = json['dynamicList'] as List
+      ..objectList = json['objectList'] as List
+      ..intList = (json['intList'] as List).map((e) => e as int).toList()
+      ..dateTimeList = (json['dateTimeList'] as List)
+          .map((e) => DateTime.parse(e as String))
+          .toList()
+      ..map = json['map'] as Map<String, dynamic>
+      ..stringStringMap =
+          new Map<String, String>.from(json['stringStringMap'] as Map)
+      ..stringIntMap = new Map<String, int>.from(json['stringIntMap'] as Map)
+      ..stringDateTimeMap = new Map<String, DateTime>.fromIterables(
+          (json['stringDateTimeMap'] as Map<String, dynamic>).keys,
+          (json['stringDateTimeMap'] as Map)
+              .values
+              .map((e) => DateTime.parse(e as String)))
+      ..crazyComplex = (json['crazyComplex'] as List)
+          .map((e) => new Map<String, Map<String, List<List<DateTime>>>>.fromIterables(
+              (e as Map<String, dynamic>).keys,
+              (e as Map)
+                  .values
+                  .map((e) => new Map<String, List<List<DateTime>>>.fromIterables((e as Map<String, dynamic>).keys, (e as Map).values.map((e) => (e as List).map((e) => (e as List).map((e) => DateTime.parse(e as String)).toList()).toList())))))
+          .toList()
+      ..val = new Map<String, bool>.from(json['val'] as Map)
+      ..writeNotNull = json['writeNotNull'] as bool
+      ..string = json[r'$string'] as String;
+
+abstract class _$KitchenSinkNonNullableSerializerMixin {
+  DateTime get dateTime;
+  Iterable<dynamic> get iterable;
+  Iterable<dynamic> get dynamicIterable;
+  Iterable<Object> get objectIterable;
+  Iterable<int> get intIterable;
+  Iterable<DateTime> get dateTimeIterable;
+  List<dynamic> get list;
+  List<dynamic> get dynamicList;
+  List<Object> get objectList;
+  List<int> get intList;
+  List<DateTime> get dateTimeList;
+  Map<dynamic, dynamic> get map;
+  Map<String, String> get stringStringMap;
+  Map<String, int> get stringIntMap;
+  Map<String, DateTime> get stringDateTimeMap;
+  List<Map<String, Map<String, List<List<DateTime>>>>> get crazyComplex;
+  Map<String, bool> get val;
+  bool get writeNotNull;
+  String get string;
+  Map<String, dynamic> toJson() {
+    var val = <String, dynamic>{};
+
+    void writeNotNull(String key, dynamic value) {
+      if (value != null) {
+        val[key] = value;
+      }
+    }
+
+    writeNotNull('dateTime', dateTime.toIso8601String());
+    writeNotNull('iterable', iterable.toList());
+    val['dynamicIterable'] = dynamicIterable.toList();
+    val['objectIterable'] = objectIterable.toList();
+    val['intIterable'] = intIterable.toList();
+    val['datetime-iterable'] =
+        dateTimeIterable.map((e) => e.toIso8601String()).toList();
+    val['list'] = list;
+    val['dynamicList'] = dynamicList;
+    val['objectList'] = objectList;
+    val['intList'] = intList;
+    writeNotNull(
+        'dateTimeList', dateTimeList.map((e) => e.toIso8601String()).toList());
+    val['map'] = map;
+    val['stringStringMap'] = stringStringMap;
+    val['stringIntMap'] = stringIntMap;
+    val['stringDateTimeMap'] = new Map<String, dynamic>.fromIterables(
+        stringDateTimeMap.keys,
+        stringDateTimeMap.values.map((e) => e.toIso8601String()));
+    writeNotNull(
+        'crazyComplex',
+        crazyComplex
+            .map((e) => new Map<String, dynamic>.fromIterables(
+                e.keys,
+                e.values.map((e) => new Map<String, dynamic>.fromIterables(
+                    e.keys,
+                    e.values.map((e) => e
+                        .map((e) => e.map((e) => e.toIso8601String()).toList())
+                        .toList())))))
+            .toList());
+    writeNotNull('val', this.val);
+    val['writeNotNull'] = this.writeNotNull;
+    val[r'$string'] = string;
+    return val;
+  }
+}