+++ json_annotation/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.2.1
+
+* `JsonSerializable` class annotation
+  * Added `nullable` field.
+  * Fixed doc comment.
+
 ## 0.2.0
 
-- Moved annotation classes for `JsonSerializable` and `JsonLiteral`.
+* Moved annotation classes for `JsonSerializable` and `JsonLiteral`.